Modify comments associations
Modified create_comments migration to implement
appropriate polymorphic relationships for
comments. Comments will always be associated
with a user and only with either a question
or an answer but never both. To access a
comments parent, you have to call
'.commentable' on the comment to return
either its answer or question parent
object as appropriate. This commit
also enables bcrypt and implements
the 'has_secure_password' method in the user
model. This commit also adds a completed
seed file. Seeding the database will
create 1000 users with the 'registered_user'
role, 5 questions asked per user, 5 questions
answered per user, 1 comment per question, and
1 comment per answer, and an admin as
the role 'admin' as prescribed in the spec.