User can vote on a question
Creates a test for a user to be able to upvote a question.
This is a feature test. Creates a model and model test for vote.
The model test for vote tests if a vote model exists, if the relationship
between question and votes exists, and if a vote can be created from a question.
Vote model tests also include a validates_inclusion_of to limit the value
of a vote to a range of -1 to 1. Model tests passing, feature still not.