Creates a new vote from a question
Builds in a button on the question show page
that passes a 1 to a newly created votes controller.
And specifies votes#create. Successfully creates a new vote
for that question, but still lacks logic to display the vote count
for the question and also to prevent a user from creating more than one question.