Complete admin delete posts feature
Adds admin ability to delete comments to bot
answers and questions from the question show
page. The button to delete lives in div with
a css id on the questioncomments and answer
comments partials. This commit adds the
detroy route for comments as well as the
method to the comments_controller.
Permission to access this method has been added
to the admin_permissions. All tests currently
passing.