Admin reactivates user
Adds methods to user model to determine
if user is a blocked user as well as
methods to deactivate and reactivate
a user. Refactored update method of
user_permissions controller with logic
to route to correct method given the
user that is passed to it. Added
a reactivate button to the view which
will route appropriately. All tests
currently passing.