Add logout
Modifies navbar with conditional that will
display a create acount button and login
button when there is no current_user.
Same conditional will display a button
to the users profile and a logout button
when there is a current_user. Adds destroy
method to sessions controller. All tests
currently passing. Behavior checked
locally in browser and is as
expected.