Complete admin blocking priveleges
Grants admin the ability to ban other users with
an admin role. Removes ability for an admin to
ban self. Adds display_block_button? and
display_unblock_button? method to application
controller; it was a significant amount of logic
cluttering up the view, so I chose to force it do
wnthe stack. Testing added to confirm new scenarios.
All tests currently passing with coverage rate of
98.83%.