Add bannable_users to admin dashboard
Adds a partial to view of user_show page that
conditionally renders if the current_user has
the role of admin. This partial renders a card
which contains the names, as links to their
profiles, of users with a reputation of <+ 10
, less than or equal to, who are not currently
banned. A method has been added to the user
model as a class method that queries the
database to return a collection of users
matching these parameters.

In order to update the reputation attribute on
the user model, a method was created in the
application controller that updates the
reputation of a user. This method is a helper method
which has been added as an after action to the
create method of the upvotes, downvotes, and
best answers controller.

All tests are currently passing. Running server
locally shows expected behavior from new feature
set.