Add UserMailer
Adds UserMailer to application. Same has methods built
to send a welcome e-mail as well as an e-mail when a user
is deactivated. Welcome e-mail is triggered when a user
is created. The blocked e-mail is triggered when the user
is deactivated. I've tested this feature in development, and
it functions appropriately.