Add email authentication
Adds authentication controller. Modifies create method
of users controller to generate a user without a role,
default guest permissions apply. When an e-mail is sent
to the user it contains a link which updates the
users permissions to be registered_user and renders
the users show page. All tests are currently passing.